testes até getAccounts

+++ db/sqlc/models.go
@@ -16,7 +16,7 @@ type Account struct {
 	Title       string       `json:"title"`
 	Type        string       `json:"type"`
 	Description string       `json:"description"`
-	Value       int32        `json:"value"`
+	Value       float32      `json:"value"`
 	Date        time.Time    `json:"date"`
 	CreatedAt   time.Time    `json:"created_at"`
 	UpdatedAt   sql.NullTime `json:"updated_at"`
